@@ -110,149 +141,113 @@
     ],
     "moduleDirectories": [
       "node_modules",
-      "build/app/node_modules"
+      "release/app/node_modules"
+    ],
+    "testPathIgnorePatterns": [
+      "release/app/dist"
     ],
     "setupFiles": [
-      "./.erb/scripts/check-build-exists.js"
+      "./.erb/scripts/check-build-exists.ts"
     ]
   },
   "devDependencies": {
-    "@babel/core": "^7.14.8",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-decorators": "^7.14.5",
-    "@babel/plugin-proposal-do-expressions": "^7.14.5",
-    "@babel/plugin-proposal-export-default-from": "^7.14.5",
-    "@babel/plugin-proposal-export-namespace-from": "^7.14.5",
-    "@babel/plugin-proposal-function-bind": "^7.14.5",
-    "@babel/plugin-proposal-function-sent": "^7.14.5",
-    "@babel/plugin-proposal-json-strings": "^7.14.5",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.14.5",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.5",
-    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
-    "@babel/plugin-proposal-pipeline-operator": "^7.14.8",
-    "@babel/plugin-proposal-throw-expressions": "^7.14.5",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@babel/plugin-transform-react-constant-elements": "^7.14.5",
-    "@babel/plugin-transform-react-inline-elements": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.15.8",
-    "@babel/preset-env": "^7.14.8",
-    "@babel/preset-react": "^7.14.5",
-    "@babel/preset-typescript": "^7.14.5",
-    "@babel/register": "^7.14.5",
-    "@pmmmwh/react-refresh-webpack-plugin": "^0.4.3",
+    "@pmmmwh/react-refresh-webpack-plugin": "0.5.4",
     "@teamsupercell/typings-for-css-modules-loader": "^2.5.1",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.7",
-    "@types/crypto-js": "^4.0.2",
-    "@types/enzyme": "^3.10.9",
-    "@types/enzyme-adapter-react-16": "^1.0.6",
-    "@types/history": "4.7.8",
-    "@types/jest": "^26.0.24",
-    "@types/node": "15.0.2",
-    "@types/react": "^17.0.9",
-    "@types/react-beforeunload": "^2.1.1",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-helmet": "^6.1.2",
-    "@types/react-router": "^5.1.14",
-    "@types/react-router-dom": "^5.1.6",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@types/command-exists": "^1.2.0",
+    "@types/jest": "^27.0.3",
+    "@types/node": "17.0.5",
+    "@types/node-static": "^0.7.7",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
     "@types/react-test-renderer": "^17.0.1",
-    "@types/stream-buffers": "^3.0.4",
-    "@types/webpack-env": "^1.16.0",
-    "@typescript-eslint/eslint-plugin": "^4.22.1",
-    "@typescript-eslint/parser": "^4.22.1",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "^26.1.0",
-    "babel-loader": "^8.2.2",
-    "babel-plugin-dev-expression": "^0.2.2",
-    "babel-plugin-direct-import": "^0.9.2",
-    "babel-plugin-import": "^1.13.3",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
-    "browserslist-config-erb": "^0.0.1",
-    "chalk": "^4.1.1",
-    "concurrently": "^6.0.2",
-    "core-js": "^3.19.0",
+    "@types/terser-webpack-plugin": "^5.0.4",
+    "@types/webpack-env": "^1.16.3",
+    "@typescript-eslint/eslint-plugin": "^5.8.1",
+    "@typescript-eslint/parser": "^5.8.1",
+    "browserslist-config-erb": "^0.0.3",
+    "chalk": "^4.1.2",
+    "concurrently": "^6.5.1",
+    "core-js": "^3.20.1",
     "cross-env": "^7.0.3",
-    "css-loader": "^5.2.4",
-    "css-minimizer-webpack-plugin": "^2.0.0",
+    "css-loader": "^6.5.1",
+    "css-minimizer-webpack-plugin": "^3.3.1",
     "detect-port": "^1.3.0",
-    "electron": "^16.0.3",
-    "electron-builder": "^22.11.1",
+    "electron": "^16.0.5",
+    "electron-builder": "22.13.1",
     "electron-devtools-installer": "^3.2.0",
-    "electron-notarize": "^1.0.0",
-    "electron-rebuild": "^2.3.5",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.6",
-    "enzyme-to-json": "^3.6.2",
-    "eslint": "^7.25.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-airbnb-typescript": "^12.3.1",
-    "eslint-config-erb": "^3.0.0",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-import-resolver-webpack": "^0.13.0",
-    "eslint-plugin-compat": "^3.9.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-jest": "^24.3.6",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-prettier": "^3.4.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "eslint-plugin-react": "^7.23.2",
-    "eslint-plugin-react-hooks": "^4.0.8",
-    "file-loader": "^6.0.0",
-    "html-webpack-plugin": "^5.3.1",
+    "electron-notarize": "^1.1.1",
+    "electron-rebuild": "^3.2.5",
+    "eslint": "^8.5.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-erb": "^4.0.3",
+    "eslint-import-resolver-typescript": "^2.5.0",
+    "eslint-import-resolver-webpack": "^0.13.2",
+    "eslint-plugin-compat": "^4.0.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jest": "^25.3.2",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react-hooks": "^4.3.0",
+    "file-loader": "^6.2.0",
+    "html-webpack-plugin": "^5.5.0",
+    "husky": "^7.0.4",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^26.1.0",
-    "lint-staged": "^10.5.4",
-    "mini-css-extract-plugin": "^1.6.0",
+    "jest": "^27.4.5",
+    "lint-staged": "^12.1.4",
+    "mini-css-extract-plugin": "^2.4.5",
     "opencollective-postinstall": "^2.0.3",
-    "prettier": "^2.2.1",
-    "prop-types": "^15.8.0",
-    "react-refresh": "^0.10.0",
+    "prettier": "^2.5.1",
+    "react-refresh": "^0.11.0",
+    "react-refresh-typescript": "^2.0.3",
     "react-test-renderer": "^17.0.2",
-    "rimraf": "^3.0.0",
-    "sass": "^1.32.12",
-    "sass-loader": "^11.0.1",
-    "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "^5.1.1",
-    "typescript": "^4.2.4",
-    "url-loader": "^4.1.0",
-    "webpack": "^5.36.2",
-    "webpack-bundle-analyzer": "^4.4.1",
-    "webpack-cli": "^4.6.0",
-    "webpack-dev-server": "^3.11.2",
-    "webpack-merge": "^5.7.3",
-    "yarn-deduplicate": "^3.1.0"
+    "rimraf": "^3.0.2",
+    "sass": "^1.45.1",
+    "sass-loader": "^12.4.0",
+    "style-loader": "^3.3.1",
+    "terser-webpack-plugin": "^5.3.0",
+    "ts-jest": "^27.1.2",
+    "ts-loader": "^9.2.6",
+    "ts-node": "^10.4.0",
+    "typescript": "^4.5.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.65.0",
+    "webpack-bundle-analyzer": "^4.5.0",
+    "webpack-cli": "^4.9.1",
+    "webpack-dev-server": "^4.7.1",
+    "webpack-merge": "^5.8.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.15.4",
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.0.1",
-    "@mui/material": "^5.0.0",
-    "@mui/styles": "^5.2.1",
+    "@emotion/react": "^11.7.1",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.5",
+    "@mui/material": "^5.2.7",
+    "@mui/styles": "^5.2.3",
     "axios": "^0.24.0",
     "command-exists": "^1.2.9",
     "crypto-js": "^4.1.1",
     "electron-debug": "^3.2.0",
-    "electron-localstorage": "^1.0.5",
-    "electron-log": "^4.3.5",
-    "electron-updater": "^4.3.8",
+    "electron-log": "^4.4.4",
+    "electron-updater": "^4.6.4",
     "form-data": "^4.0.0",
-    "form-data-encoder": "^1.7.1",
-    "formdata-node": "^4.3.1",
-    "history": "4.x.x",
+    "history": "^5.2.0",
+    "image-size": "^1.0.0",
     "node-static": "^0.7.11",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
     "react-router-dom": "^5.2.0",
-    "react-toastify": "^8.0.3",
-    "regenerator-runtime": "^0.13.5"
+    "react-toastify": "^8.1.0",
+    "regenerator-runtime": "^0.13.9"
   },
   "devEngines": {
-    "node": ">=10.x",
-    "npm": ">=6.x",
-    "yarn": ">=1.21.3"
+    "node": ">=14.x",
+    "npm": ">=7.x"
+  },
+  "collective": {
+    "url": "https://opencollective.com/electron-react-boilerplate-594"
   },
   "browserslist": [],
   "prettier": {
