@@ -1,8 +1,11 @@
 module.exports = {
   extends: 'erb',
   rules: {
-    // temporary hack related to IDE not resolving correct package.json
+    // A temporary hack related to IDE not resolving correct package.json
     'import/no-extraneous-dependencies': 'off',
+    'import/no-unresolved': 'error',
+    // Since React 17 and typescript 4.1 you can safely disable the rule
+    'react/react-in-jsx-scope': 'off',
     'react/self-closing-comp': 'off',
     'react/sort-comp': 'off',
     'react/prefer-stateless-function': 'off',
@@ -51,8 +54,9 @@ module.exports = {
       // See https://github.com/benmosher/eslint-plugin-import/issues/1396#issuecomment-575727774 for line below
       node: {},
       webpack: {
-        config: require.resolve('./.erb/configs/webpack.config.eslint.js'),
+        config: require.resolve('./.erb/configs/webpack.config.eslint.ts'),
       },
+      typescript: {},
     },
     'import/parsers': {
       '@typescript-eslint/parser': ['.ts', '.tsx'],
