@@ -1,4 +1,4 @@
-/* eslint @typescript-eslint/no-explicit-any: off */
+type Styles = Record<string, string>;
 
 declare module '*.svg' {
   const content: string;
@@ -12,5 +12,20 @@ declare module '*.png' {
 
 declare module '*.jpg' {
   const content: string;
+  export default content;
+}
+
+declare module '*.scss' {
+  const content: Styles;
+  export default content;
+}
+
+declare module '*.sass' {
+  const content: Styles;
+  export default content;
+}
+
+declare module '*.css' {
+  const content: Styles;
   export default content;
 }
