@@ -1,9 +1,10 @@
 import { createBrowserHistory } from 'history';
+
 import Credentials from 'structs/Credentials';
 import { UIEvents } from 'renderer/UIEvents';
 import GLOBALS from 'shared/globals';
 import UserData from 'structs/UserData';
-import { DebugRendererHandler } from './DebugLogger';
+import { DebugRendererHandler } from 'shared/DebugLogger';
 import { IElectronRendererWindow } from 'types/types';
 
 export const history = createBrowserHistory();
